requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
anthropic>=0.18.0
Pillow>=10.0.0
